rug-hunter-bot_backup_20251114_070934/backend/requirements_complete.txt
# ================================
# RUG HUNTER BOT v4.0 - REQUIREMENTS
# ================================

# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Web3 & Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1
solana==0.30.2
base58==2.1.1

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning (Optionnel - commentez si vous n'utilisez pas le ML)
scikit-learn==1.3.2
# tensorflow==2.15.0  # Très lourd, décommentez si besoin
# xgboost==2.0.3
# lightgbm==4.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.0

# Terminal Display
rich==13.7.0
colorama==0.4.6
tabulate==0.9.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Cryptography
cryptography==41.0.7
pycryptodome==3.19.0

# Async & Concurrency
aiodns==3.1.1

# Logging
loguru==0.7.2

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# JSON & Data
orjson==3.9.10
msgpack==1.0.7

# Notifications (Optional)
python-telegram-bot==20.7
discord-webhook==1.3.1

# Utils
tenacity==8.2.3
cachetools<5.0.0,>=4.2.2  # Compatible avec solana
typing-extensions==4.9.0